ARROW-6266: [Java] Resolve the ambiguous method overload in RangeEqualsVisitor
In RangeEqualsVisitor, there are overload methods for both super class and sub class. This will lead to unexpected behavior.

For example, if we call RangeEqualsVisitor#visit(v), where v is a fixed width vector, the method actually called may be visit(ValueVector), which is unexpected.

In general, in the visitor pattern, it is not a good idea to support method overload for both super class and sub-class as parameters.

Closes #5100 from liyafan82/fly_0816_visit and squashes the following commits:

6b5e5c548 <liyafan82>  Remove the general visit mothod
f7ed538e2 <liyafan82>  Resolve the ambiguous method overload in RangeEqualsVisitor

Authored-by: liyafan82 <fan_li_ya@foxmail.com>
Signed-off-by: Pindikura Ravindra <ravindra@dremio.com>