ARROW-6247: [Java] Provide a common interface for float4 and float8 vectors
We want to provide an interface for floating point vectors (float4 & float8). This interface will make it convenient for many operations on a vector. With this interface, the client code will be greatly simplified, with many branches/switch removed.

The design is similar to BaseIntVector (the interface for all integer vectors). We provide 3 methods for setting & getting floating point values:

         setWithPossibleTruncate

         setSafeWithPossibleTruncate

         getValueAsDouble

Closes #5132 from liyafan82/fly_0820_float and squashes the following commits:

984d53851 <liyafan82>  Provide a common interface for float4 and float8 vectors

Authored-by: liyafan82 <fan_li_ya@foxmail.com>
Signed-off-by: Micah Kornfield <emkornfield@gmail.com>