ARROW-6416: [Python] Improve API & documentation regarding chunksizes
The current API and its documentation suggests files written with `RecordBatchFileWriter.write_table(..., chunksize)` will be written with a fixed chunk size when in fact the chunksize parameter is an upper bound on the size of the chunks to be written.

In my opinion the parameter `chunksize` should therefore be renamed `max_chunksize` to avoid confusion and reflect its true purpose.

This would also improve naming consistency in the code base, since in the C++ implementation this parameter is already named `max_chunksize` in `cpp/source/arrow/ipc/writer.cc`.

Similarly, the parameter should be renamed in `Table.to_batches(chunksize)`.

Closes #5254 from ARF1/ARROW-6416 and squashes the following commits:

1c8847538 <ARF1> suggestion for graceful API deprecation
569bf9b98 <ARF1> ARROW-6416:  Improve API & documentation regarding chunksizes

Authored-by: ARF1 <ARF1@github>
Signed-off-by: Wes McKinney <wesm+git@apache.org>