ARROW-6379: [C++] Write no IPC buffer metadata for NullType
Since we have no integration tests for NullType we have never had to address this issue.

The C++ implementation has been writing 2 `Buffer` Flatbuffer struct values with length 0 for NullType. Rather than having dummy/placeholder Buffer I think it is more consistent to write no metadata for this type.

The downside of this change is that any persisted record batches from pre 0.15.0 will not be readable. However, we have not made any forward compatibility guarantees about this so I am OK breaking a few eggs on this one.

Closes #5287 from wesm/ARROW-6379 and squashes the following commits:

638a43fe8 <Wes McKinney> Write no IPC buffer metadata for NullType

Authored-by: Wes McKinney <wesm+git@apache.org>
Signed-off-by: Wes McKinney <wesm+git@apache.org>