ARROW-7019: [Java] Improve the performance of loading validity buffers
At the receiver side of flighting, loading validity buffer is an important operation, as each vector has a validity buffer.

For non-nullable vectors, the current implementation of loading the validity buffer is inefficient. We improve the performance of this operation by efficiently setting the bits of a memory region to 1.

Benchmark results show that the changes leads to a 35% performance improvement:

Before:
BitVectorHelperBenchmarks.loadValidityBufferAllOne avgt 5 748.916 ± 23.290 ns/op

After:
BitVectorHelperBenchmarks.loadValidityBufferAllOne avgt 5 487.352 ± 15.046 ns/op

Closes #5751 from liyafan82/fly_1029_one and squashes the following commits:

cfde3c798 <liyafan82>  Improve the performance of loading validity buffers

Authored-by: liyafan82 <fan_li_ya@foxmail.com>
Signed-off-by: Neal Richardson <neal.p.richardson@gmail.com>