ARROW-6888: [Java] Support copy operation for vector value comparators
In this issue, we provide copy operations for vector value comparators. This operation creates another comparator with the same type and comparison logic.

This feature is useful in multi-threading scenarios where multiple threads uses the comparator to perform their own task. In this scenario, we have no way of making sure the compare method is thread safe. So a safe way is to create a new comparator for each thread. The copy operation will support this.

An immediate application of this is the parallel searcher for ordering semantics.

Closes #5659 from liyafan82/fly_1015_cp and squashes the following commits:

740342a68 <liyafan82>  Support CompositeVectorComparator
a3c703b72 <liyafan82>  Change to createNew method
0b783bf4e <liyafan82>  Implement cloneable interface
a6ba74d2f <liyafan82>  Support copy operation for vector value comparators

Authored-by: liyafan82 <fan_li_ya@foxmail.com>
Signed-off-by: Micah Kornfield <emkornfield@gmail.com>