ARROW-7168: [Python] Respect the specified dictionary type for pd.Categorical conversion
https://issues.apache.org/jira/browse/ARROW-7168

This change ensures that if you specify a `type` in `pa.array`, we ensure the output actually has this type when converting to dictionary array (as we also do for other types).

The PR now implements this change, but we might want to do this with a deprecation first, as this can break people's code.

Closes #5866 from jorisvandenbossche/ARROW-7168-categorical-specified-type and squashes the following commits:

39ff8e82c <Joris Van den Bossche> more python 2
e4dbb2c4f <Joris Van den Bossche> try fix python 2
003e6532b <Joris Van den Bossche> for now use deprecation warnings instead of error
bfb82372b <Joris Van den Bossche> additional tests
3535a5699 <Joris Van den Bossche> ARROW-7168:  Respect the specified dictionary type when converting pd.Categorical

Authored-by: Joris Van den Bossche <jorisvandenbossche@gmail.com>
Signed-off-by: Antoine Pitrou <antoine@python.org>