ARROW-6863: [Java] Provide parallel searcher
For scenarios where the vector is large and the a low response time is required, we need to search the vector in parallel to improve the responsiveness.

This issue tries to provide a parallel searcher for the equality semantics (the support for ordering semantics is not ready yet, as we need a way to distribute the comparator).

The implementation is based on multi-threading.

Closes #5631 from liyafan82/fly_1011_par and squashes the following commits:

25e86f479 <liyafan82>  Refine Javadoc
5b21cc3cd <liyafan82>  Provide benchmark for the parallel searcher
e735bcc24 <liyafan82>  Provide parallel searcher

Authored-by: liyafan82 <fan_li_ya@foxmail.com>
Signed-off-by: Neal Richardson <neal.p.richardson@gmail.com>