ARROW-7998: [C++][Plasma] Make Seal requests synchronous
When handling a `Seal` request to create an object and make it visible to other clients, the plasma store does not wait until the seal is complete before responding to the requesting client. This makes the interface hard to use, since the client is not guaranteed that the object is visible yet and would have to use an additional IPC round-trip to determine when the object is ready.

This improvement would require the plasma store to wait until the object has been created before responding to the client.

Closes #6529 from stephanie-wang/sync-seal and squashes the following commits:

cb3867e4f <Stephanie Wang> Make Seal synchronous

Authored-by: Stephanie Wang <swang@cs.berkeley.edu>
Signed-off-by: Philipp Moritz <pcmoritz@gmail.com>