ARROW-8009: [Java] Fix the hash code methods for BitVector
The current hash code methods of BitVector are based on implementations in BaseFixedWidthVector, which rely on the type width of the vector.
For BitVector, the type width is 0, so the underlying data is not actually used when computing the hash code. That means, the hash code will always be 0, no matter if the underlying data is null or not, and no matter if the underlying bit is 0 or 1.

We fix this by overriding the methods in BitVector.

Closes #6543 from liyafan82/fly_0305_bit and squashes the following commits:

a38b52795 <liyafan82>  Fix the hash code methods for BitVector

Authored-by: liyafan82 <fan_li_ya@foxmail.com>
Signed-off-by: Micah Kornfield <emkornfield@gmail.com>