ARROW-8030: [Plasma] Uniform comments style
The comments in the plasma are a mixture of '@params' and '\params'. The reviewers required me to unify the style when I was trying to add windows support. I think it would be better to address it using a different PR. This is a blocking issue of a few refactoring and the windows support pr, so hopefully the review process is fast.

Closes #6559 from suquark/fix_comments and squashes the following commits:

f9858e210 <Siyuan> uniform comments style

Authored-by: Siyuan <suquark@gmail.com>
Signed-off-by: Sutou Kouhei <kou@clear-code.com>