PARQUET-1786: [C++] Improve ByteStreamSplit decoder using SSE2
Since the data is already likely in the cache, memory accesses
are sufficiently fast to use SIMD intrinsics to speed-up decoding.
This patch adds a decode path which is only taken when Arrow is
compiled with SSE2 and higher support.
The BYTE_STREAM_SPLIT round-trip test size is purposely changed to
a value non-divisible by 4 and 8 to guarantee that the incomplete
SIMD suffix is correctly handled by the scalar decode loop.

Closes #6679 from martinradev/byte_stream_split_submit

Lead-authored-by: Martin Radev <martin.b.radev@gmail.com>
Co-authored-by: Antoine Pitrou <antoine@python.org>
Signed-off-by: Antoine Pitrou <antoine@python.org>