ARROW-7891: [C++][GLib][Python][R] Make uniform use of check_metadata=false default. Add Py/R/GLib bindings for RecordBatch::Equals with check_metadata
It seems empirically that having `check_metadata = true` as the default can be a nuisance, and checking the metadata will take place in fewer cases than checking the field names, types, and nullability for equality.

I carried through these changes in Python and R, and added bindings for RecordBatch::Equals with the check_metadata argument.

This is an API change, but IMHO a good one.

Closes #6830 from wesm/ARROW-7891

Lead-authored-by: Wes McKinney <wesm+git@apache.org>
Co-authored-by: Sutou Kouhei <kou@clear-code.com>
Co-authored-by: Wes McKinney <wesm@users.noreply.github.com>
Signed-off-by: Neal Richardson <neal.p.richardson@gmail.com>