ARROW-8508: [Rust] FixedSizeListArray improper offset for value
Fixes [ARROW-8508](https://issues.apache.org/jira/browse/ARROW-8508)

Fixed size list arrays sourced with a non-zero offset of their
child data were respecting this offset in the `value_offset` method,
but not in the `value` method. This would cause nested fixed list
arrays that should have looked like this:

```
[
  [ [0, 1] ],
  [ [2, 3], [4, 5] ]
]
```

...to behave like this when looking at the values using the `value()`
method (like how the debug output does).

```
[
  [ [0, 1] ],
  [ [0, 1], [2, 3] ],
]
```

This is different to how ListArray would do things (which respect the offset in both methods). This PR makes the change. Additionally, it adds a failing test case for this, as well as a passing test case for similar cases, including on ListArray.

Closes #7006 from markhildreth/ARROW-8508

Authored-by: Mark Hildreth <mark.k.hildreth@gmail.com>
Signed-off-by: Andy Grove <andygrove@nvidia.com>