ARROW-8783: [Rust] [DataFusion] Add ParquetScan and CsvScan variants in LogicalPlan
This PR makes it possible to build a logical plan that reads CSV or Parquet data sources without the need to create an execution context and register data sources.

This makes it easier for other crates to leverage the DataFusion logical plan builder and query optimizer and then execute the query by other means. It also simplifies direct use of DataFusion in some cases.

Closes #7164 from andygrove/ARROW-8783

Authored-by: Andy Grove <andygrove73@gmail.com>
Signed-off-by: Andy Grove <andygrove@nvidia.com>