ARROW-6052: [C++] Split up arrow/array.h/cc into multiple files under arrow/array/, move ArrayData to separate header, make ArrayData::dictionary ArrayData
One meaningful change in this patch was to change `ArrayData::dictionary` to be `shared_ptr<ArrayData>`, which was not difficult.

Otherwise, this allows some files that only need ArrayData or just Array to not include a bunch of unnecessary header code. The increase in number of lines of code is pretty much all due to license headers.

I did a modest amount of IWYU cleaning, there are plenty of still usages of "arrow/array.h" that might be better replaced by more specific headers.

As part of this patch I disabled `build/include_what_you_use` in cpplint per ARROW-8994. If we're going to spend time doing IWYU cleaning (which IMHO is overall worthwhile) we should rely on its output rather than having IWYU checks by two different tools.

Closes #7310 from wesm/cpp-split-array-h

Authored-by: Wes McKinney <wesm+git@apache.org>
Signed-off-by: Wes McKinney <wesm+git@apache.org>