ARROW-8927: [C++] Support dictionary memo in CUDA IPC ReadRecordBatch functions
Adds support for dictionary memo in `ReadRecordBatch` for CUDA IPC, passing the memo directly to the `ipc::ReadRecordBatch` function. This allows one to handle record batches with dictionaries, though it might be worth adding some tests around using `SerializeSchema` as well, so we can have the complete CPU + GPU IPC path in place -- let me know and I'd be happy to do so.

Closes #7263 from alexbaden/ARROW-8927

Lead-authored-by: Alex Baden <alex.baden@gmail.com>
Co-authored-by: Antoine Pitrou <antoine@python.org>
Signed-off-by: Antoine Pitrou <antoine@python.org>