ARROW-8402: [Java] Support ValidateFull methods in Java
In this PR, we provide the validateFull functionalities, and refactor the validate logic. We separate the validation logic into 3 visitors:

1. Type validation: validate vector type information (O(1) time complexity).
2. Buffer validation: validate buffer count/size, etc. (O(1) time complexity).
3. Data validation: validate content of data buffers. (O(n) time complexity).

For the `validate` method, we perform validations 1 and 2, with a total time complexity of O(1).
For the `validateFull` method, we perform validations 1, 2, and 3, with a total time complexity of O(n).

Closes #7248 from liyafan82/fly_0515_full

Authored-by: liyafan82 <fan_li_ya@foxmail.com>
Signed-off-by: Micah Kornfield <emkornfield@gmail.com>